# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Roster HR Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /designations:
    get:
      operationId: listDesignations
      tags: [designations]
      summary: List all designations
      responses:
        "200":
          description: List of designations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Designation"
    post:
      operationId: createDesignation
      tags: [designations]
      summary: Create designation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesignationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Designation"
  /designations/{id}:
    patch:
      operationId: updateDesignation
      tags: [designations]
      summary: Update designation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesignationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Designation"
    delete:
      operationId: deleteDesignation
      tags: [designations]
      summary: Delete designation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Employees ──────────────────────────────────────────────────────────────
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees with optional filters
      parameters:
        - name: department
          in: query
          schema:
            type: ["integer", "null"]
        - name: designation
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeList"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Soft-delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/direct-reports:
    get:
      operationId: getDirectReports
      tags: [employees]
      summary: Get direct reports for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of direct reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeSummary"
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload a document record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Announcements ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Designation:
      type: object
      required: [id, name, departmentId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: ["string", "null"]
        createdAt:
          type: string
    DesignationInput:
      type: object
      required: [name, departmentId]
      properties:
        name:
          type: string
          minLength: 1
        departmentId:
          type: integer
    DesignationUpdate:
      type: object
      properties:
        name:
          type: string
        departmentId:
          type: integer

    # ── Employees ─────────────────────────────────────────────────────────────
    EmployeeSummary:
      type: object
      required: [id, fullName, employeeCode, status]
      properties:
        id:
          type: integer
        fullName:
          type: string
        employeeCode:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        departmentName:
          type: ["string", "null"]
        designationName:
          type: ["string", "null"]
        status:
          type: string
    Employee:
      type: object
      required: [id, fullName, employeeCode, email, status, hireDate, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        employeeCode:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        designationId:
          type: ["integer", "null"]
        designationName:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, on_leave, terminated]
        employmentType:
          type: string
          enum: [full_time, part_time, contract, intern]
        hireDate:
          type: string
        salary:
          type: ["number", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        createdAt:
          type: string
    EmployeeList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    EmployeeInput:
      type: object
      required: [fullName, email, hireDate]
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        designationId:
          type: ["integer", "null"]
        managerId:
          type: ["integer", "null"]
        status:
          type: string
          default: active
        employmentType:
          type: string
          default: full_time
        hireDate:
          type: string
        salary:
          type: ["number", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
    EmployeeUpdate:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        designationId:
          type: ["integer", "null"]
        managerId:
          type: ["integer", "null"]
        status:
          type: string
        employmentType:
          type: string
        hireDate:
          type: string
        salary:
          type: ["number", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
...
    Document:
      type: object
      required: [id, name, type, uploadedById, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: ["integer", "null"]
        employeeName:
          type: ["string", "null"]
        name:
          type: string
        type:
          type: string
          enum: [contract, id_proof, certificate, policy, other]
        fileUrl:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        uploadedById:
          type: integer
        isCompanyPolicy:
          type: boolean
        createdAt:
          type: string
    DocumentInput:
      type: object
      required: [name, type]
      properties:
        employeeId:
          type: ["integer", "null"]
        name:
          type: string
          minLength: 1
        type:
          type: string
        fileUrl:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        isCompanyPolicy:
          type: boolean
          default: false

    # ── Announcements ─────────────────────────────────────────────────────────
...